Added utils/resources.py and changed calls to util.read_file*
