+++ qutebrowser/config/configdata.py
@@ -30,7 +30,7 @@ import functools
 import dataclasses
 
 from qutebrowser.config import configtypes
-from qutebrowser.utils import usertypes, qtutils, utils
+from qutebrowser.utils import usertypes, qtutils, utils, resources
 from qutebrowser.misc import debugcachestats
 
 DATA = cast(Mapping[str, 'Option'], None)
@@ -272,4 +272,4 @@ def is_valid_prefix(prefix: str) -> bool:
 def init() -> None:
     """Initialize configdata from the YAML file."""
     global DATA, MIGRATIONS
-    DATA, MIGRATIONS = _read_yaml(utils.read_file('config/configdata.yml'))
+    DATA, MIGRATIONS = _read_yaml(resources.read_file('config/configdata.yml'))
