+++ qutebrowser/utils/utils.py
@@ -195,105 +195,6 @@ def compact_text(text: str, elidelength: int = None) -> str:
         out = elide(out, elidelength)
     return out
 
-
-def _resource_path(filename: str) -> pathlib.Path:
-    """Get a pathlib.Path object for a resource."""
-    assert not posixpath.isabs(filename), filename
-    assert os.path.pardir not in filename.split(posixpath.sep), filename
-
-    if hasattr(sys, 'frozen'):
-        # For PyInstaller, where we can't store resource files in a qutebrowser/ folder
-        # because the executable is already named "qutebrowser" (at least on macOS).
-        return pathlib.Path(sys.executable).parent / filename
-
-    return importlib_resources.files(qutebrowser) / filename
-
-
-@contextlib.contextmanager
-def _resource_keyerror_workaround() -> Iterator[None]:
-    """Re-raise KeyErrors as FileNotFoundErrors.
-
-    WORKAROUND for zipfile.Path resources raising KeyError when a file was notfound:
-    https://bugs.python.org/issue43063
-
-    Only needed for Python 3.8 and 3.9.
-    """
-    try:
-        yield
-    except KeyError as e:
-        raise FileNotFoundError(str(e))
-
-
-def _glob_resources(
-    resource_path: pathlib.Path,
-    subdir: str,
-    ext: str,
-) -> Iterable[str]:
-    """Find resources with the given extension.
-
-    Yields a resource name like "html/log.html" (as string).
-    """
-    assert '*' not in ext, ext
-    assert ext.startswith('.'), ext
-    path = resource_path / subdir
-
-    if isinstance(resource_path, pathlib.Path):
-        for full_path in path.glob(f'*{ext}'):  # . is contained in ext
-            yield full_path.relative_to(resource_path).as_posix()
-    else:  # zipfile.Path or importlib_resources compat object
-        # Unfortunately, we can't tell mypy about resource_path being of type
-        # Union[pathlib.Path, zipfile.Path] because we set "python_version = 3.6" in
-        # .mypy.ini, but the zipfiel stubs (correctly) only declare zipfile.Path with
-        # Python 3.8...
-        assert path.is_dir(), path  # type: ignore[unreachable]
-        for subpath in path.iterdir():
-            if subpath.name.endswith(ext):
-                yield posixpath.join(subdir, subpath.name)
-
-
-def preload_resources() -> None:
-    """Load resource files into the cache."""
-    resource_path = _resource_path('')
-    for subdir, ext in [
-            ('html', '.html'),
-            ('javascript', '.js'),
-            ('javascript/quirks', '.js'),
-    ]:
-        for name in _glob_resources(resource_path, subdir, ext):
-            _resource_cache[name] = read_file(name)
-
-
-def read_file(filename: str) -> str:
-    """Get the contents of a file contained with qutebrowser.
-
-    Args:
-        filename: The filename to open as string.
-
-    Return:
-        The file contents as string.
-    """
-    if filename in _resource_cache:
-        return _resource_cache[filename]
-
-    path = _resource_path(filename)
-    with _resource_keyerror_workaround():
-        return path.read_text(encoding='utf-8')
-
-
-def read_file_binary(filename: str) -> bytes:
-    """Get the contents of a binary file contained with qutebrowser.
-
-    Args:
-        filename: The filename to open as string.
-
-    Return:
-        The file contents as a bytes object.
-    """
-    path = _resource_path(filename)
-    with _resource_keyerror_workaround():
-        return path.read_bytes()
-
-
 def parse_version(version: str) -> VersionNumber:
     """Parse a version string."""
     ver, _suffix = QVersionNumber.fromString(version)
